Remove code that's no longer used by SSLyze
