+++ nassl/_nassl.pyi
@@ -4,7 +4,6 @@ SSL_CTX: Any
 SSL: Any
 BIO: Any
 X509: Any
-X509_STORE_CTX: Any
 OCSP_RESPONSE: Any
 OpenSSLError: Any
 SslError: Any
