+++ setup.py
@@ -141,7 +141,6 @@ MODERN_NASSL_EXT_SETUP.update(
         ],
     }
 )
-MODERN_NASSL_EXT_SETUP["sources"].append("nassl/_nassl/nassl_X509_STORE_CTX.c")  # API only available in modern nassl
 
 
 if CURRENT_PLATFORM in [
