--- tests/X509_STORE_CTX_test.py
@@ -1,55 +1,0 @@
-import pytest
-
-from nassl._nassl import X509, X509_STORE_CTX
-
-
-@pytest.fixture
-def certificate_as_x509() -> X509:
-    pem_cert = """-----BEGIN CERTIFICATE-----
-MIIDdTCCAl2gAwIBAgILBAAAAAABFUtaw5QwDQYJKoZIhvcNAQEFBQAwVzELMAkG
-A1UEBhMCQkUxGTAXBgNVBAoTEEdsb2JhbFNpZ24gbnYtc2ExEDAOBgNVBAsTB1Jv
-b3QgQ0ExGzAZBgNVBAMTEkdsb2JhbFNpZ24gUm9vdCBDQTAeFw05ODA5MDExMjAw
-MDBaFw0yODAxMjgxMjAwMDBaMFcxCzAJBgNVBAYTAkJFMRkwFwYDVQQKExBHbG9i
-YWxTaWduIG52LXNhMRAwDgYDVQQLEwdSb290IENBMRswGQYDVQQDExJHbG9iYWxT
-aWduIFJvb3QgQ0EwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQDaDuaZ
-jc6j40+Kfvvxi4Mla+pIH/EqsLmVEQS98GPR4mdmzxzdzxtIK+6NiY6arymAZavp
-xy0Sy6scTHAHoT0KMM0VjU/43dSMUBUc71DuxC73/OlS8pF94G3VNTCOXkNz8kHp
-1Wrjsok6Vjk4bwY8iGlbKk3Fp1S4bInMm/k8yuX9ifUSPJJ4ltbcdG6TRGHRjcdG
-snUOhugZitVtbNV4FpWi6cgKOOvyJBNPc1STE4U6G7weNLWLBYy5d4ux2x8gkasJ
-U26Qzns3dLlwR5EiUWMWea6xrkEmCMgZK9FGqkjWZCrXgzT/LCrBbBlDSgeF59N8
-9iFo7+ryUp9/k5DPAgMBAAGjQjBAMA4GA1UdDwEB/wQEAwIBBjAPBgNVHRMBAf8E
-BTADAQH/MB0GA1UdDgQWBBRge2YaRQ2XyolQL30EzTSo//z9SzANBgkqhkiG9w0B
-AQUFAAOCAQEA1nPnfE920I2/7LqivjTFKDK1fPxsnCwrvQmeU79rXqoRSLblCKOz
-yj1hTdNGCbM+w6DjY1Ub8rrvrTnhQ7k4o+YviiY776BQVvnGCv04zcQLcFGUl5gE
-38NflNUVyRRBnMRddWQVDf9VMOyGj/8N7yy5Y0b2qvzfvGn9LhJIZJrglfCm7ymP
-AbEVtQwdpf5pLGkkeB6zpxxxYu7KyJesF12KwvhHhm4qxFYxldBniYUr+WymXUad
-DKqC5JlR3XC321Y9YeRq4VzW9v493kHMB65jUr9TU/Qr6cf9tveCX4XSQRjbgbME
-HMUfpIBvFSDJ3gyICh3WZlXi/EjJKSZp4A==
------END CERTIFICATE-----"""
-    return X509(pem_cert)
-
-
-class TestX509_STORE_CTX:
-    def test_set0_trusted_stack(self, certificate_as_x509: X509) -> None:
-        ctx = X509_STORE_CTX()
-        ctx.set0_trusted_stack([certificate_as_x509, certificate_as_x509])
-
-        # When calling it a second time it fails
-        with pytest.raises(ValueError):
-            ctx.set0_trusted_stack([certificate_as_x509, certificate_as_x509])
-
-    def test_set0_untrusted(self, certificate_as_x509: X509) -> None:
-        ctx = X509_STORE_CTX()
-        ctx.set0_untrusted([certificate_as_x509, certificate_as_x509])
-
-        # When calling it a second time it fails
-        with pytest.raises(ValueError):
-            ctx.set0_untrusted([certificate_as_x509, certificate_as_x509])
-
-    def test_set_cert(self, certificate_as_x509: X509) -> None:
-        ctx = X509_STORE_CTX()
-        ctx.set_cert(certificate_as_x509)
-
-        # When calling it a second time it fails
-        with pytest.raises(ValueError):
-            ctx.set_cert(certificate_as_x509)
